Phase 2 step 12: Migrate AccountsStrategiesMixin (37 methods)
